Mirror test and src dir structure
